analizzatore_foto_video/analizzatore video 2/requirements.txt.txt
opencv-python==4.9.0.80
mediapipe==0.10.11
pillow==10.3.0
numpy==1.26.4